perf(channels): Phase 1 of #2815 - tiered wait + NetworkDetector-gated yield
Replaces the unconditional task::run(250, SYSTEM) -> vTaskDelay(>=1 tick)
yield inside ChannelManager::waitForCondition and IChannelDriver::waitForCondition
with a tiered structure:

  Tier 1 - instant non-blocking condition check (free fast path)
  Tier 2 - bounded microsecond spin (FASTLED_WAIT_SPIN_BUDGET_US = 250 default)
  Tier 3 - cooperator-friendly deep yield, gated on whether a radio is
           actually active via fl::NetworkDetector::isAnyNetworkActive()

When no WiFi/Ethernet/BT is up (the #2420 reporter's case), the deep yield
collapses to a non-tick-flooring fast yield - same path that the existing
FL_IS_ESP_32P4 carve-out used for ESP32-P4. That carve-out is removed and
replaced by the runtime NetworkDetector gate; P4 still gets the fast path
via NetworkDetector's weak-symbol fallback (no WiFi/BT components linked).

Closes #2818. See #2815 for the full design rationale and the per-phase
plan (Phase 2 adds the per-driver waitDone() ISR fast path).

Changes:
- New shared helper: src/fl/channels/detail/wait_yield.{h,cpp.hpp}
- Promoted fl::NetworkDetector from drivers/rmt/rmt_5/ to platforms/esp/32/util/
  so non-RMT5 ESP32 builds (PARLIO-only ESP32-P4, etc.) can use it. The old
  rmt_5/network_detector.{h,cpp.hpp} paths are forwarding shims.
- Public API: FastLED.setWaitSpinBudgetUs(N) runtime setter +
  FASTLED_WAIT_SPIN_BUDGET_US compile-time default (per Public Settings
  Pattern - new global setters go on CFastLED).
- ChannelManager::waitForCondition and IChannelDriver::waitForCondition
  now delegate to fl::detail::tieredWait.
