+++ src/FastLED.cpp.hpp
@@ -11,6 +11,7 @@
 #include "fl/channels/manager.h"
 #include "fl/system/trace.h"
 #include "fl/channels/driver.h"  // for IChannelDriver
+#include "fl/channels/detail/wait_yield.h"  // for tiered-wait spin-budget setters (#2818)
 #include "fl/system/delay.h"  // for delayMicroseconds
 #include "fl/system/sketch_macros.h"
 #if SKETCH_HAS_LARGE_MEMORY
@@ -595,6 +596,16 @@ void CFastLED::wait() {
 bool CFastLED::wait(fl::u32 timeout_ms) {
 	fl::ChannelManager& manager = fl::channelManager();
 	return manager.waitForReady(timeout_ms);
+}
+
+// Tiered-wait spin-budget shims — actual storage lives in
+// fl/channels/detail/wait_yield.cpp.hpp (#2818).
+void CFastLED::_setWaitSpinBudgetUs(fl::u32 budget_us) FL_NOEXCEPT {
+	fl::detail::setWaitSpinBudgetUs(budget_us);
+}
+
+fl::u32 CFastLED::_getWaitSpinBudgetUs() FL_NOEXCEPT {
+	return fl::detail::getWaitSpinBudgetUs();
 }
 
 // ============================================================================
