+++ src/fl/channels/detail/_build.cpp.hpp
@@ -1,4 +1,8 @@
 /// @file _build.cpp.hpp
 /// @brief Unity build header for fl/channels/detail/ directory
 
+// begin current directory includes
+#include "fl/channels/detail/wait_yield.cpp.hpp"
+
+// begin sub directory includes
 #include "fl/channels/detail/validation/_build.cpp.hpp"
