+++ src/fl/channels/manager.cpp.hpp
@@ -2,6 +2,7 @@
 /// @brief Implementation of unified channel bus manager
 
 #include "fl/channels/manager.h"
+#include "fl/channels/detail/wait_yield.h"
 #include "fl/stl/singleton.h"
 #include "fl/log/log.h"
 #include "fl/log/log.h"
@@ -362,29 +363,15 @@ fl::shared_ptr<IChannelDriver> ChannelManager::selectDriverForChannel(const Chan
 
 template<typename Condition>
 bool ChannelManager::waitForCondition(Condition condition, u32 timeoutMs) {
-    const u32 startTime = timeoutMs > 0 ? millis() : 0;
-
-    while (!condition()) {
-        // Check timeout if specified
-        if (timeoutMs > 0 && (millis() - startTime >= timeoutMs)) {
-            FL_ERROR("ChannelManager: Timeout occurred while waiting for condition");
-            return false;  // Timeout occurred
-        }
-
-#if defined(FL_IS_ESP_32P4)
-        // ESP32-P4 has no network stack (no WiFi, no Bluetooth, no lwIP),
-        // so the deep-yield rationale from #2254 doesn't apply. taskYIELD()
-        // avoids the 1-tick (≥1 ms at CONFIG_FREERTOS_HZ=1000) floor that
-        // adds ~8 ms per frame in PARLIO multi-strip workloads (#2493).
-        task::run(0, task::ExecFlags::SYSTEM);
-#else
-        // OS yield only — keeps WiFi/lwIP alive without pumping
-        // tasks or coroutines during frame transitions (re-entrancy risk).
-        task::run(250, task::ExecFlags::SYSTEM);
-#endif
+    // Tiered: instant → spin (FASTLED_WAIT_SPIN_BUDGET_US default 250 µs)
+    // → cooperator yield with NetworkDetector gating. See #2815/#2818.
+    // The previous FL_IS_ESP_32P4 carve-out is subsumed by NetworkDetector
+    // returning "no radio" on P4 via weak-symbol fallback.
+    const bool ok = fl::detail::tieredWait(condition, timeoutMs);
+    if (!ok) {
+        FL_ERROR("ChannelManager: Timeout occurred while waiting for condition");
     }
-
-    return true;  // Condition met
+    return ok;
 }
 
 IChannelDriver::DriverState ChannelManager::poll() {
