+++ src/platforms/esp/32/_build.cpp.hpp
@@ -23,3 +23,4 @@
 #include "platforms/esp/32/drivers/_build.cpp.hpp"
 #include "platforms/esp/32/interrupts/_build.cpp.hpp"
 #include "platforms/esp/32/ota/_build.cpp.hpp"
+#include "platforms/esp/32/util/_build.cpp.hpp"
