+++ src/platforms/esp/32/util/_build.cpp.hpp
@@ -1,0 +1,6 @@
+// IWYU pragma: private
+
+/// @file _build.hpp
+/// @brief Unity build header for platforms/esp/32/util/ directory
+
+#include "platforms/esp/32/util/network_detector.cpp.hpp"
